mbta-api/requirements.txt
beautifulsoup4==4.11.1
certifi==2024.12.14
distlib==0.3.9
filelock==3.16.1
numpy==1.23.4
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
soupsieve==2.3.2.post1
urllib3==1.26.12
virtualenv==20.29.0
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.9.0.post0
